refactor!: relocate fetch.ts to types
