+++ website-frontend/src/lib/directus.ts
@@ -2,7 +2,7 @@ import { createDirectus, rest, authentication } from '@directus/sdk';
 import { EMAIL, PASSWORD } from '$env/static/private';
 import { PUBLIC_APIURL } from '$env/static/public';
 import { type Schema } from './models/schema';
-import type { Fetch } from './fetch';
+import type { Fetch } from './types/fetch';
 
 async function getDirectusInstance(fetch: Fetch | null) {
 	const options = fetch ? { globals: { fetch } } : {};
